fix: Gets the tutorial to compile
There were some instances didn't need to be there anymore. This is what
I had to tidy up to get things to compile when I was playing with it.

Also put it into a separate library to compile it nicely, which you
probably don't want. More a heads up that some instances need nuking. :)