Change DMapV's implementation to use `:.:` instead of `Compose`.
The latter has a lot of baggage regarding instances. For example, to get
Eq for a Compose structure one will rely on the Eq1 class. It's very
frequently the case that we do not want to use this class because it
cannot be easily generated for GADTs.

Vessel is already using `constraints-extras` extensively to wrangle
instances for such GADTs and :.: is more neutral about how instances are
fulfilled based on the constituent pieces. For example compare the two:

`Eq (f (g p)) => Eq ((f :.: g) p)`

`(Eq1 f, Eq1 g, Eq a) => Eq (Compose f g a)`

The former will be easier to write by hand with constraints-extras's
`ArgDict` which can be generated for you.