Give `Vessel` and `DMapV` more polymorphic kinds.
In particular for `Vessel`, its second parameter can be `x -> *` for any
`x` and therefore its first parameter has kind `(x -> *) -> *) -> *`.

For `DMapV` both the key and value functor parameters can use any kind
as index, so their kinds are `x -> *` for any `x`.

This is a minor enhancement but a useful one. The indexes don't actually
play a part in any of the data manipulation and in some use cases there
is a convenient kind to index over that is not *.